+++ lib/objects/voucher_children/unlimited_percentage_voucher_with_end_time.dart
@@ -8,7 +8,7 @@ import '../../generated/l10n.dart';
 
 class UnlimitedPercentageVoucherWithEndTime extends Voucher
     implements EndTimeInterface, PercentageInterface {
-  double _maximumDiscountValue;
+  int _maximumDiscountValue;
   DateTime _endTime;
 
   UnlimitedPercentageVoucherWithEndTime({
@@ -26,7 +26,7 @@ class UnlimitedPercentageVoucherWithEndTime extends Voucher
     super.hasEndTime = true,
     super.isLimited = false,
     required DateTime endTime,
-    required double maximumDiscountValue,
+    required int maximumDiscountValue,
   })  : _endTime = endTime,
         _maximumDiscountValue = maximumDiscountValue;
 
@@ -36,9 +36,9 @@ class UnlimitedPercentageVoucherWithEndTime extends Voucher
   set endTime(DateTime value) => _endTime = value;
 
   @override
-  double get maximumDiscountValue => _maximumDiscountValue;
+  int get maximumDiscountValue => _maximumDiscountValue;
   @override
-  set maximumDiscountValue(double value) => _maximumDiscountValue = value;
+  set maximumDiscountValue(int value) => _maximumDiscountValue = value;
 
   @override
   void updateVoucher({
@@ -46,14 +46,14 @@ class UnlimitedPercentageVoucherWithEndTime extends Voucher
     String? voucherName,
     DateTime? startTime,
     double? discountValue,
-    double? minimumPurchase,
+    int? minimumPurchase,
     int? maxUsagePerPerson,
     bool? isVisible,
     bool? isEnabled,
     String? enDescription,
     String? viDescription,
     DateTime? endTime,
-    double? maximumDiscountValue,
+    int? maximumDiscountValue,
   }) {
     super.updateVoucher(
       voucherID: voucherID,
@@ -155,7 +155,7 @@ class UnlimitedPercentageVoucherWithEndTime extends Voucher
     String? voucherName,
     DateTime? startTime,
     double? discountValue,
-    double? minimumPurchase,
+    int? minimumPurchase,
     int? maxUsagePerPerson,
     bool? isVisible,
     bool? isEnabled,
@@ -165,7 +165,7 @@ class UnlimitedPercentageVoucherWithEndTime extends Voucher
     bool? hasEndTime,
     bool? isLimited,
     DateTime? endTime,
-    double? maximumDiscountValue,
+    int? maximumDiscountValue,
   }) {
     return UnlimitedPercentageVoucherWithEndTime(
       voucherID: voucherID ?? this.voucherID,
