+++ lib/objects/voucher_children/unlimited_percentage_voucher_without_end_time.dart
@@ -7,7 +7,7 @@ import '../../generated/l10n.dart';
 
 class UnlimitedPercentageVoucherWithoutEndTime extends Voucher
     implements PercentageInterface {
-  double _maximumDiscountValue;
+  int _maximumDiscountValue;
 
   UnlimitedPercentageVoucherWithoutEndTime({
     super.voucherID,
@@ -23,13 +23,13 @@ class UnlimitedPercentageVoucherWithoutEndTime extends Voucher
     super.isPercentage = false,
     super.hasEndTime = true,
     super.isLimited = false,
-    required double maximumDiscountValue,
+    required int maximumDiscountValue,
   }) : _maximumDiscountValue = maximumDiscountValue;
 
   @override
-  double get maximumDiscountValue => _maximumDiscountValue;
+  int get maximumDiscountValue => _maximumDiscountValue;
   @override
-  set maximumDiscountValue(double value) => _maximumDiscountValue = value;
+  set maximumDiscountValue(int value) => _maximumDiscountValue = value;
 
   @override
   void updateVoucher({
@@ -37,14 +37,14 @@ class UnlimitedPercentageVoucherWithoutEndTime extends Voucher
     String? voucherName,
     DateTime? startTime,
     double? discountValue,
-    double? minimumPurchase,
+    int? minimumPurchase,
     int? maxUsagePerPerson,
     bool? isVisible,
     bool? isEnabled,
     String? enDescription,
     String? viDescription,
     DateTime? endTime,
-    double? maximumDiscountValue,
+    int? maximumDiscountValue,
   }) {
     super.updateVoucher(
       voucherID: voucherID,
@@ -139,7 +139,7 @@ class UnlimitedPercentageVoucherWithoutEndTime extends Voucher
     String? voucherName,
     DateTime? startTime,
     double? discountValue,
-    double? minimumPurchase,
+    int? minimumPurchase,
     int? maxUsagePerPerson,
     bool? isVisible,
     bool? isEnabled,
@@ -149,7 +149,7 @@ class UnlimitedPercentageVoucherWithoutEndTime extends Voucher
     bool? hasEndTime,
     bool? isLimited,
     DateTime? endTime,
-    double? maximumDiscountValue,
+    int? maximumDiscountValue,
   }) {
     return UnlimitedPercentageVoucherWithoutEndTime(
       voucherID: voucherID ?? this.voucherID,
