+++ lib/objects/voucher_related/voucher.dart
@@ -7,7 +7,7 @@ abstract class Voucher {
   String voucherName;
   DateTime startTime;
   double discountValue;
-  double minimumPurchase;
+  int minimumPurchase;
   int maxUsagePerPerson;
   bool isVisible;
   bool isEnabled;
@@ -39,7 +39,7 @@ abstract class Voucher {
     String? voucherID,
     String? voucherName,
     double? discountValue,
-    double? minimumPurchase,
+    int? minimumPurchase,
     int? maxUsagePerPerson,
     DateTime? startTime,
 
@@ -65,7 +65,7 @@ abstract class Voucher {
     String? voucherName,
     DateTime? startTime,
     double? discountValue,
-    double? minimumPurchase,
+    int? minimumPurchase,
     int? maxUsagePerPerson,
     bool? isVisible,
     bool? isEnabled,
