+++ lib/screens/voucher/add_voucher/add_voucher_view.dart
@@ -192,7 +192,7 @@ class _AddVoucherScreen extends State<AddVoucherScreen> {
                                     ),
                                     const SizedBox(width: 16),
                                     Expanded(
-                                      child: buildInputWidget<double>(
+                                      child: buildInputWidget<int>(
                                         S.of(context).minimumPurchase,
                                         minimumPurchaseController,
                                         state.voucherArgument?.minimumPurchase,
@@ -309,7 +309,7 @@ class _AddVoucherScreen extends State<AddVoucherScreen> {
                                 if (state.voucherArgument?.isPercentage == true)
                                   Padding(
                                     padding: const EdgeInsets.only(top: 16.0),
-                                    child: buildInputWidget<double>(
+                                    child: buildInputWidget<int>(
                                       S.of(context).maximumDiscountValue,
                                       maximumDiscountValueController,
                                       state.voucherArgument
