+++ lib/screens/voucher/edit_voucher/edit_voucher_view.dart
@@ -219,14 +219,12 @@ class _EditVoucherScreen extends State<EditVoucherScreen> {
                                     ),
                                     const SizedBox(width: 16),
                                     Expanded(
-                                      child: buildInputWidget<double>(
+                                      child: buildInputWidget<int>(
                                         S.of(context).minimumPurchase,
                                         minimumPurchaseController,
                                         state.voucherArgument?.minimumPurchase,
                                             (value) {
-                                          cubit.updateVoucherArgument(
-                                              state.voucherArgument!.copyWith(
-                                                  minimumPurchase: value));
+                                          cubit.updateVoucherArgument(state.voucherArgument!.copyWith(minimumPurchase: value));
                                         },
                                       ),
                                     ),
@@ -336,15 +334,13 @@ class _EditVoucherScreen extends State<EditVoucherScreen> {
                                 if (state.voucherArgument?.isPercentage == true)
                                   Padding(
                                     padding: const EdgeInsets.only(top: 16.0),
-                                    child: buildInputWidget<double>(
+                                    child: buildInputWidget<int>(
                                       S.of(context).maximumDiscountValue,
                                       maximumDiscountValueController,
                                       state.voucherArgument
                                           ?.maximumDiscountValue,
                                           (value) {
-                                        cubit.updateVoucherArgument(
-                                            state.voucherArgument!.copyWith(
-                                                maximumDiscountValue: value));
+                                        cubit.updateVoucherArgument(state.voucherArgument!.copyWith(maximumDiscountValue: value));
                                       },
                                     ),
                                   ),
